#000 refactor the return type to long because the value of position is 50 exceeds the int type

+++ src/main/java/cc/xpbootcamp/warmup/fibonacci/Fibonacci.java
@@ -1,10 +1,10 @@
 package cc.xpbootcamp.warmup.fibonacci;
 
 public class Fibonacci {
-    public int calculate(int i) {
+    public long calculate(int i) {
         if (i > 2) {
             return calculate(i - 1) + calculate(i - 2);
         }
-        return 1;
+        return 1L;
     }
 }
